Database.Persist.Class.DeleteCascade

Description

Deprecated: The DeleteCascade module is deprecated. You can now set cascade behavior directly on entities in the quasiquoter.

Documentation

class (PersistStoreWrite backend, PersistEntity record, BaseBackend backend ~ PersistEntityBackend record) => DeleteCascade record backend where

For combinations of backends and entities that support cascade-deletion. “Cascade-deletion” means that entries that depend on other entries to be deleted will be deleted as well.

Methods

deleteCascade :: MonadIO m => Key record -> ReaderT backend m ()

Perform cascade-deletion of single database entry.

deleteCascadeWhere :: forall record backend m. (MonadIO m, DeleteCascade record backend, PersistQueryWrite backend) => [Filter record] -> ReaderT backend m ()

Cascade-deletion of entries satisfying given filters.