Database.Persist.FieldDef

Description

Since: 2.13.0.0

The FieldDef type

data FieldDef

A FieldDef represents the inormation that persistent knows about a field of a datatype. This includes information used to parse the field out of the database and what the field corresponds to.

Setters

setFieldAttrs :: [FieldAttr] -> FieldDef -> FieldDef

Replace the FieldDef FieldAttr with the new list.

Since: 2.13.0.0

overFieldAttrs :: ([FieldAttr] -> [FieldAttr]) -> FieldDef -> FieldDef

Modify the list of field attributes.

Since: 2.13.0.0

addFieldAttr :: FieldAttr -> FieldDef -> FieldDef

Add an attribute to the list of field attributes.

Since: 2.13.0.0

Helpers

isFieldNullable :: FieldDef -> IsNullable

Check if the field definition is nullable

Since: 2.13.0.0

isFieldMaybe :: FieldDef -> Bool

Check if the field is `Maybe a`

Since: 2.13.0.0

isFieldNotGenerated :: FieldDef -> Bool

isHaskellField :: FieldDef -> Bool

Returns True if the FieldDef does not have a MigrationOnly or SafeToRemove flag from the QuasiQuoter.

Since: 2.13.0.0

FieldCascade

data FieldCascade

This datatype describes how a foreign reference field cascades deletes or updates.

This type is used in both parsing the model definitions and performing migrations. A Nothing in either of the field values means that the user has not specified a CascadeAction. An unspecified CascadeAction is defaulted to Restrict when doing migrations.

Since: 2.11.0

Constructors

FieldCascade

Fields

fcOnUpdate :: !(Maybe CascadeAction)

fcOnDelete :: !(Maybe CascadeAction)

renderFieldCascade :: FieldCascade -> Text

Renders a FieldCascade value such that it can be used in SQL migrations.

Since: 2.11.0

renderCascadeAction :: CascadeAction -> Text

Render a CascadeAction to Text such that it can be used in a SQL command.

Since: 2.11.0

noCascade :: FieldCascade

A FieldCascade that does nothing.

Since: 2.11.0

data CascadeAction

An action that might happen on a deletion or update on a foreign key change.

Since: 2.11.0

Constructors

Cascade

Restrict

SetNull

SetDefault