Database.Persist.Names

Description

This module contains types and functions for working with and disambiguating database and Haskell names.

Since: 2.13.0.0

Documentation

class DatabaseName a where

Convenience operations for working with '-NameDB' types.

Since: 2.12.0.0

Methods

escapeWith :: (Text -> str) -> a -> str

newtype FieldNameDB

An EntityNameDB represents the datastore-side name that persistent will use for an entity.

Since: 2.12.0.0

Constructors

FieldNameDB

Fields

unFieldNameDB :: Text

newtype FieldNameHS

A FieldNameHS represents the Haskell-side name that persistent will use for a field.

Since: 2.12.0.0

Constructors

FieldNameHS

Fields

unFieldNameHS :: Text

newtype EntityNameHS

An EntityNameHS represents the Haskell-side name that persistent will use for an entity.

Since: 2.12.0.0

Constructors

EntityNameHS

Fields

unEntityNameHS :: Text

newtype EntityNameDB

An EntityNameDB represents the datastore-side name that persistent will use for an entity.

Since: 2.12.0.0

Constructors

EntityNameDB

Fields

unEntityNameDB :: Text

newtype ConstraintNameDB

A ConstraintNameDB represents the datastore-side name that persistent will use for a constraint.

Since: 2.12.0.0

Constructors

ConstraintNameDB

Fields

unConstraintNameDB :: Text

newtype ConstraintNameHS

An ConstraintNameHS represents the Haskell-side name that persistent will use for a constraint.

Since: 2.12.0.0

Constructors

ConstraintNameHS

Fields

unConstraintNameHS :: Text