Database.Persist.Query.Internal

Documentation

class PersistStore backend m => PersistQuery backend m where

Methods

update :: PersistEntity val => Key backend val -> [Update val] -> backend m ()

Update individual fields on a specific record.

updateGet :: PersistEntity val => Key backend val -> [Update val] -> backend m val

Update individual fields on a specific record, and retrieve the updated value from the database.

Note that this function will throw an exception if the given key is not found in the database.

updateWhere :: PersistEntity val => [Filter val] -> [Update val] -> backend m ()

Update individual fields on any record matching the given criterion.

deleteWhere :: PersistEntity val => [Filter val] -> backend m ()

Delete all records matching the given criterion.

selectSource :: (PersistEntity val, PersistEntityBackend val ~ backend) => [Filter val] -> [SelectOpt val] -> Source (ResourceT (backend m)) (Entity val)

Get all records matching the given criterion in the specified order. Returns also the identifiers.

selectFirst :: (PersistEntity val, PersistEntityBackend val ~ backend) => [Filter val] -> [SelectOpt val] -> backend m (Maybe (Entity val))

get just the first record for the criterion

selectKeys :: PersistEntity val => [Filter val] -> [SelectOpt val] -> Source (ResourceT (backend m)) (Key backend val)

Get the Keys of all records matching the given criterion.

count :: PersistEntity val => [Filter val] -> backend m Int

The total number of records fulfilling the given criterion.

selectList :: (PersistEntity val, PersistQuery backend m, PersistEntityBackend val ~ backend) => [Filter val] -> [SelectOpt val] -> backend m [Entity val]

Call selectSource but return the result as a list.

selectKeysList :: (PersistEntity val, PersistQuery b m, PersistEntityBackend val ~ b) => [Filter val] -> [SelectOpt val] -> b m [Key b val]

Call selectKeys but return the result as a list.

data SelectOpt v

Constructors

forall typ . Asc (EntityField v typ)

forall typ . Desc (EntityField v typ)

OffsetBy Int

LimitTo Int

limitOffsetOrder :: PersistEntity val => [SelectOpt val] -> (Int, Int, [SelectOpt val])

data Filter v

Filters which are available for select, updateWhere and deleteWhere. Each filter constructor specifies the field being filtered on, the type of comparison applied (equals, not equals, etc) and the argument for the comparison.

Constructors

forall typ . PersistField typ => Filter

Fields

filterField :: EntityField v typ

filterValue :: Either typ [typ]

filterFilter :: PersistFilter

FilterAnd [Filter v]
convenient for internal use, not needed for the API

FilterOr [Filter v]

data PersistUpdate

Constructors

Assign

Add

Subtract

Multiply

Divide

data Update v

Constructors

forall typ . PersistField typ => Update

Fields

updateField :: EntityField v typ

updateValue :: typ

updateUpdate :: PersistUpdate

updateFieldDef :: PersistEntity v => Update v -> FieldDef

deleteCascadeWhere :: (DeleteCascade a backend m, PersistQuery backend m) => [Filter a] -> backend m ()

Exceptions

data UpdateGetException

Constructors

KeyNotFound String