Database.Persist.SqlBackend

Description

This module contains types and information necessary for a SQL database. Database support libraries, like persistent-postgresql, will be responsible for constructing these values.

The type and construction

data SqlBackend

A SqlBackend represents a handle or connection to a database. It contains functions and values that allow databases to have more optimized implementations, as well as references that benefit performance and sharing.

Instead of using the SqlBackend constructor directly, use the mkSqlBackend function.

A SqlBackend is *not* thread-safe. You should not assume that a SqlBackend can be shared among threads and run concurrent queries. This *will* result in problems. Instead, you should create a Pool SqlBackend, known as a ConnectionPool, and pass that around in multi-threaded applications.

To run actions in the persistent library, you should use the runSqlConn function. If you're using a multithreaded application, use the runSqlPool function.

mkSqlBackend :: MkSqlBackendArgs -> SqlBackend

A function for creating a value of the SqlBackend type. You should prefer to use this instead of the constructor for SqlBackend, because default values for this will be provided for new fields on the record when new functionality is added.

Since: 2.13.0.0

data MkSqlBackendArgs

This type shares many of the same field names as the SqlBackend type. It's useful for library authors to use this when migrating from using the SqlBackend constructor directly to the mkSqlBackend function.

This type will only contain required fields for constructing a SqlBackend. For fields that aren't present on this record, you'll want to use the various set functions or

Since: 2.13.0.0

Constructors

MkSqlBackendArgs

Fields

connPrepare :: Text -> IO Statement

This function should prepare a Statement in the target database, which should allow for efficient query reuse.

connInsertSql :: EntityDef -> [PersistValue] -> InsertSqlResult

This function generates the SQL and values necessary for performing an insert against the database.

connStmtMap :: IORef (Map Text Statement)

A reference to the cache of statements. Statements are keyed by the Text queries that generated them.

connClose :: IO ()

Close the underlying connection.

connMigrateSql :: [EntityDef] -> (Text -> IO Statement) -> EntityDef -> IO (Either [Text] [(Bool, Text)])

This function returns the migrations required to include the EntityDef parameter in the [EntityDef] database. This might include creating a new table if the entity is not present, or altering an existing table if it is.

connBegin :: (Text -> IO Statement) -> Maybe IsolationLevel -> IO ()

A function to begin a transaction for the underlying database.

connCommit :: (Text -> IO Statement) -> IO ()

A function to commit a transaction to the underlying database.

connRollback :: (Text -> IO Statement) -> IO ()

A function to roll back a transaction on the underlying database.

connEscapeFieldName :: FieldNameDB -> Text

A function to extract and escape the name of the column corresponding to the provided field.

Since: 2.12.0.0

connEscapeTableName :: EntityDef -> Text

A function to extract and escape the name of the table corresponding to the provided entity. PostgreSQL uses this to support schemas.

Since: 2.12.0.0

connEscapeRawName :: Text -> Text

A function to escape raw DB identifiers. MySQL uses backticks, while PostgreSQL uses quotes, and so on.

Since: 2.12.0.0

connNoLimit :: Text

connRDBMS :: Text

A tag displaying what database the SqlBackend is for. Can be used to differentiate features in downstream libraries for different database backends.

connLimitOffset :: (Int, Int) -> Text -> Text

Attach a 'LIMIT/OFFSET' clause to a SQL query. Note that LIMIT/OFFSET is problematic for performance, and indexed range queries are the superior way to offer pagination.

connLogFunc :: LogFunc

A log function for the SqlBackend to use.

data SqlBackendHooks

emptySqlBackendHooks :: SqlBackendHooks

Utilities

The functions exported here are a bit more general than the record accessors. The easiest way to use them is to provide the SqlBackend directly to the function. However, you can also use them in a ReaderT context, and you can even use them with any backend type tht has a BackendCompatible SqlBackend backend instance.

SqlBackend Getters

getRDBMS :: (BackendCompatible SqlBackend backend, MonadReader backend m) => m Text

Get a tag displaying what database the SqlBackend is for. Can be used to differentiate features in downstream libraries for different database backends. @since 2.13.3.0

getEscapedFieldName :: (BackendCompatible SqlBackend backend, MonadReader backend m) => FieldNameDB -> m Text

This function can be used directly with a SqlBackend to escape a FieldNameDB.

let conn :: SqlBackend
getEscapedFieldName (FieldNameDB "asdf") conn

Alternatively, you can use it in a ReaderT SqlBackend context, like SqlPersistT:

query :: SqlPersistM Text
query = do
    field <- getEscapedFieldName (FieldNameDB "asdf")
    pure field

Since: 2.13.0.0

getEscapedRawName :: (BackendCompatible SqlBackend backend, MonadReader backend m) => Text -> m Text

This function can be used directly with a SqlBackend to escape a raw Text.

let conn :: SqlBackend
getEscapedRawName (FieldNameDB "asdf") conn

Alternatively, you can use it in a ReaderT SqlBackend context, like SqlPersistT:

query :: SqlPersistM Text
query = do
    field <- getEscapedRawName (FieldNameDB "asdf")
    pure field

Since: 2.13.0.0

getEscapeRawNameFunction :: (BackendCompatible SqlBackend backend, MonadReader backend m) => m (Text -> Text)

Return the function for escaping a raw name.

Since: 2.13.0.0

getConnLimitOffset

Arguments

:: (BackendCompatible SqlBackend backend, MonadReader backend m)

=> (Int, Int)
The (LIMIT, OFFSET) to put on the query.

-> Text
The SQL query that the LIMIT/OFFSET clause will be attached to.

-> m Text

Decorate the given SQL query with the (LIMIT, OFFSET) specified.

Since: 2.13.0.0

getConnUpsertSql :: (BackendCompatible SqlBackend backend, MonadReader backend m) => m (Maybe (EntityDef -> NonEmpty (FieldNameHS, FieldNameDB) -> Text -> Text))

Retrieve the function for generating an upsert statement, if the backend supports it.

Since: 2.13.0.0

getConnVault :: (BackendCompatible SqlBackend backend, MonadReader backend m) => m Vault

Retrieve the vault from the provided database backend.

Since: 2.13.3.0

getConnHooks :: (BackendCompatible SqlBackend backend, MonadReader backend m) => m SqlBackendHooks

Retrieve instrumentation hooks from the provided database backend.

Since: 2.13.3.0

SqlBackend Setters

setConnMaxParams :: Int -> SqlBackend -> SqlBackend

Set the maximum parameters that may be issued in a given SQL query. This should be used only if the database backend have this limitation.

Since: 2.13.0.0

setConnRepsertManySql :: (EntityDef -> Int -> Text) -> SqlBackend -> SqlBackend

Set the connRepsertManySql field on the SqlBackend. This should only be set by the database backend library. If this is not set, a slow default will be used.

Since: 2.13.0.0

setConnInsertManySql :: (EntityDef -> [[PersistValue]] -> InsertSqlResult) -> SqlBackend -> SqlBackend

Set the connInsertManySql field on the SqlBackend. This should only be used by the database backend library to provide an efficient implementation of a bulk insert function. If this is not set, a slow default will be used.

Since: 2.13.0.0

setConnUpsertSql :: (EntityDef -> NonEmpty (FieldNameHS, FieldNameDB) -> Text -> Text) -> SqlBackend -> SqlBackend

Set the connUpsertSql field on the SqlBackend. This should only be used by the database backend library to provide an efficient implementation of a bulk insert function. If this is not set, a slow default will be used.

Since: 2.13.0.0

setConnPutManySql :: (EntityDef -> Int -> Text) -> SqlBackend -> SqlBackend

Set the 'connPutManySql field on the SqlBackend. This should only be used by the database backend library to provide an efficient implementation of a bulk insert function. If this is not set, a slow default will be used.

Since: 2.13.0.0

setConnVault :: Vault -> SqlBackend -> SqlBackend

Set the vault on the provided database backend.

Since: 2.13.0

modifyConnVault :: (Vault -> Vault) -> SqlBackend -> SqlBackend

Modify the vault on the provided database backend.

Since: 2.13.0

setConnHooks :: SqlBackendHooks -> SqlBackend -> SqlBackend

Set hooks on the provided database backend.

Since: 2.13.0

SqlBackendHooks