Database.Persist.SqlBackend.Internal.Statement

Documentation

data Statement

A Statement is a representation of a database query that has been prepared and stored on the server side.

Constructors

Statement

Fields

stmtFinalize :: IO ()

stmtReset :: IO ()

stmtExecute :: [PersistValue] -> IO Int64

stmtQuery :: forall m. MonadIO m => [PersistValue] -> Acquire (ConduitM () [PersistValue] m ())