Database.Persist.Store

Description

API for database actions. The API deals with fields and entities. In SQL, a field corresponds to a column, and should be a single non-composite value. An entity corresponds to a SQL table, so an entity is a collection of fields.

Documentation

data PersistValue

A raw value which can be stored in any backend and can be marshalled to and from a PersistField.

Constructors

PersistText Text

PersistByteString ByteString

PersistInt64 Int64

PersistDouble Double

PersistBool Bool

PersistDay Day

PersistTimeOfDay TimeOfDay

PersistUTCTime UTCTime

PersistZonedTime ZT

PersistNull

PersistList [PersistValue]

PersistMap [(Text, PersistValue)]

PersistObjectId ByteString
intended especially for MongoDB backend

data SqlType

A SQL data type. Naming attempts to reflect the underlying Haskell datatypes, eg SqlString instead of SqlVarchar. Different SQL databases may have different translations for these types.

Constructors

SqlString

SqlInt32

SqlInt64

SqlReal

SqlBool

SqlDay

SqlTime

SqlDayTime

SqlDayTimeZoned

SqlBlob

SqlOther Text
a backend-specific name

class PersistField a where

A value which can be marshalled to and from a PersistValue.

Methods

toPersistValue :: a -> PersistValue

fromPersistValue :: PersistValue -> Either Text a

sqlType :: a -> SqlType

isNullable :: a -> Bool

class PersistEntity val where

A single database entity. For example, if writing a blog application, a blog entry would be an entry, containing fields such as title and content.

Associated Types

data EntityField val :: * -> *

Parameters: val and datatype of the field

type PersistEntityBackend val :: (* -> *) -> * -> *

data Unique val :: ((* -> *) -> * -> *) -> *

Unique keys in existence on this entity.

Methods

persistFieldDef :: EntityField val typ -> FieldDef

entityDef :: val -> EntityDef

toPersistFields :: val -> [SomePersistField]

fromPersistValues :: [PersistValue] -> Either Text val

halfDefined :: val

persistUniqueToFieldNames :: Unique val backend -> [(HaskellName, DBName)]

persistUniqueToValues :: Unique val backend -> [PersistValue]

persistUniqueKeys :: val -> [Unique val backend]

persistIdField :: EntityField val (Key (PersistEntityBackend val) val)

class (MonadBaseControl IO m, MonadBaseControl IO (backend m)) => PersistStore backend m where

Methods

insert :: PersistEntity val => val -> backend m (Key backend val)

Create a new record in the database, returning an automatically created key (in SQL an auto-increment id).

insertKey :: PersistEntity val => Key backend val -> val -> backend m ()

Create a new record in the database using the given key.

repsert :: PersistEntity val => Key backend val -> val -> backend m ()

Put the record in the database with the given key. Unlike replace, if a record with the given key does not exist then a new record will be inserted.

replace :: PersistEntity val => Key backend val -> val -> backend m ()

Replace the record in the database with the given key. Note that the result is undefined if such record does not exist, so you must use insertKey or repsert in these cases.

delete :: PersistEntity val => Key backend val -> backend m ()

Delete a specific record by identifier. Does nothing if record does not exist.

get :: PersistEntity val => Key backend val -> backend m (Maybe val)

Get a record by identifier, if available.

class PersistStore backend m => PersistUnique backend m where

Methods

getBy :: (PersistEntityBackend val ~ backend, PersistEntity val) => Unique val backend -> backend m (Maybe (Entity val))

Get a record by unique key, if available. Returns also the identifier.

deleteBy :: PersistEntity val => Unique val backend -> backend m ()

Delete a specific record by unique key. Does nothing if no record matches.

insertUnique :: (backend ~ PersistEntityBackend val, PersistEntity val) => val -> backend m (Maybe (Key backend val))

Like insert, but returns Nothing when the record couldn't be inserted because of a uniqueness constraint.

data PersistFilter

Constructors

Eq

Ne

Gt

Lt

Ge

Le

In

NotIn

BackendSpecificFilter Text

data SomePersistField

Constructors

forall a . PersistField a => SomePersistField a

newtype ZT

Avoid orphan instances.

Constructors

ZT ZonedTime

insertBy :: (PersistEntity v, PersistStore backend m, PersistUnique backend m, backend ~ PersistEntityBackend v) => v -> backend m (Either (Entity v) (Key backend v))

Insert a value, checking for conflicts with any unique constraints. If a duplicate exists in the database, it is returned as Left. Otherwise, the new Key is returned as Right.

getByValue :: (PersistEntity v, PersistUnique backend m, PersistEntityBackend v ~ backend) => v -> backend m (Maybe (Entity v))

A modification of getBy, which takes the PersistEntity itself instead of a Unique value. Returns a value matching one of the unique keys. This function makes the most sense on entities with a single Unique constructor.

getJust :: (PersistStore backend m, PersistEntity val, Show (Key backend val)) => Key backend val -> backend m val

Same as get, but for a non-null (not Maybe) foreign key Unsafe unless your database is enforcing that the foreign key is valid

belongsTo :: (PersistStore backend m, PersistEntity ent1, PersistEntity ent2) => (ent1 -> Maybe (Key backend ent2)) -> ent1 -> backend m (Maybe ent2)

belongsToJust :: (PersistStore backend m, PersistEntity ent1, PersistEntity ent2) => (ent1 -> Key backend ent2) -> ent1 -> backend m ent2

same as belongsTo, but uses getJust and therefore is similarly unsafe

checkUnique :: (PersistEntityBackend val ~ backend, PersistEntity val, PersistUnique backend m) => val -> backend m Bool

Check whether there are any conflicts for unique keys with this entity and existing entities in the database.

Returns True if the entity would be unique, and could thus safely be inserted; returns False on a conflict.

class PersistEntity a => DeleteCascade a backend m where

Methods

deleteCascade :: Key backend a -> backend m ()

data PersistException

Constructors

PersistError Text
Generic Exception

PersistMarshalError Text

PersistInvalidField Text

PersistForeignConstraintUnmet Text

PersistMongoDBError Text

PersistMongoDBUnsupported Text

newtype Key backend entity

Constructors

Key

Fields

unKey :: PersistValue

data Entity entity

Datatype that represents an entity, with both its key and its Haskell representation.

When using the an SQL-based backend (such as SQLite or PostgreSQL), an Entity may take any number of columns depending on how many fields it has. In order to reconstruct your entity on the Haskell side, persistent needs all of your entity columns and in the right order. Note that you don't need to worry about this when using persistent's API since everything is handled correctly behind the scenes.

However, if you want to issue a raw SQL command that returns an Entity, then you have to be careful with the column order. While you could use SELECT Entity.* WHERE ... and that would work most of the time, there are times when the order of the columns on your database is different from the order that persistent expects (for example, if you add a new field in the middle of you entity definition and then use the migration code -- persistent will expect the column to be in the middle, but your DBMS will put it as the last column). So, instead of using a query like the one above, you may use rawSql (from the Database.Persist.GenericSql module) with its /entity selection placeholder/ (a double question mark ??). Using rawSql the query above must be written as SELECT ?? WHERE ... Then rawSql will replace ?? with the list of all columns that we need from your entity in the right order. If your query returns two entities (i.e. (Entity backend a, Entity backend b)), then you must you use SELECT ??, ?? WHERE ..., and so on.

Constructors

Entity

Fields

entityKey :: Key (PersistEntityBackend entity) entity

entityVal :: entity

Helpers

getPersistMap :: PersistValue -> Either Text [(Text, PersistValue)]

listToJSON :: [PersistValue] -> Text

mapToJSON :: [(Text, PersistValue)] -> Text

Config

class PersistConfig c where

Represents a value containing all the configuration options for a specific backend. This abstraction makes it easier to write code that can easily swap backends.

Associated Types

type PersistConfigBackend c :: (* -> *) -> * -> *

type PersistConfigPool c

Methods

loadConfig :: Value -> Parser c

Load the config settings from a Value, most likely taken from a YAML config file.

applyEnv :: c -> IO c

Modify the config settings based on environment variables.

createPoolConfig :: c -> IO (PersistConfigPool c)

Create a new connection pool based on the given config settings.

runPool :: (MonadBaseControl IO m, MonadIO m) => c -> PersistConfigBackend c m a -> PersistConfigPool c -> m a

Run a database action by taking a connection from the pool.