Database.Persist.Sql.Types.Internal

Documentation

class HasPersistBackend backend where

Class which allows the plucking of a BaseBackend backend from some larger type. For example, instance HasPersistBackend (SqlReadBackend, Int) where type BaseBackend (SqlReadBackend, Int) = SqlBackend persistBackend = unSqlReadBackend . fst

Associated Types

type BaseBackend backend

Methods

persistBackend :: backend -> BaseBackend backend

class HasPersistBackend backend => IsPersistBackend backend where

Class which witnesses that backend is essentially the same as BaseBackend backend. That is, they're isomorphic and backend is just some wrapper over BaseBackend backend.

Methods

mkPersistBackend :: BaseBackend backend -> backend

This function is how we actually construct and tag a backend as having read or write capabilities. It should be used carefully and only when actually constructing a backend. Careless use allows us to accidentally run a write query against a read-only database.

data SqlReadBackend

An SQL backend which can only handle read queries

data SqlWriteBackend

An SQL backend which can handle read or write queries

readToUnknown :: Monad m => ReaderT SqlReadBackend m a -> ReaderT SqlBackend m a

Useful for running a read query against a backend with unknown capabilities.

readToWrite :: Monad m => ReaderT SqlReadBackend m a -> ReaderT SqlWriteBackend m a

Useful for running a read query against a backend with read and write capabilities.

writeToUnknown :: Monad m => ReaderT SqlWriteBackend m a -> ReaderT SqlBackend m a

Useful for running a write query against an untagged backend with unknown capabilities.

type LogFunc = Loc -> LogSource -> LogLevel -> LogStr -> IO ()

data InsertSqlResult

Constructors

ISRSingle Text

ISRInsertGet Text Text

ISRManyKeys Text [PersistValue]

data Statement

Constructors

Statement

Fields

stmtFinalize :: IO ()

stmtReset :: IO ()

stmtExecute :: [PersistValue] -> IO Int64

stmtQuery :: forall m. MonadIO m => [PersistValue] -> Acquire (ConduitM () [PersistValue] m ())

data IsolationLevel

Please refer to the documentation for the database in question for a full overview of the semantics of the varying isloation levels

Constructors

ReadUncommitted

ReadCommitted

RepeatableRead

Serializable

makeIsolationLevelStatement :: (Monoid s, IsString s) => IsolationLevel -> s

data SqlBackend

Constructors

SqlBackend

Fields

connPrepare :: Text -> IO Statement

connInsertSql :: EntityDef -> [PersistValue] -> InsertSqlResult

table name, column names, id name, either 1 or 2 statements to run

connInsertManySql :: Maybe (EntityDef -> [[PersistValue]] -> InsertSqlResult)

SQL for inserting many rows and returning their primary keys, for backends that support this functioanlity. If Nothing, rows will be inserted one-at-a-time using connInsertSql.

connUpsertSql :: Maybe (EntityDef -> Text -> Text)

Some databases support performing UPSERT _and_ RETURN entity in a single call.

This field when set will be used to generate the UPSERT+RETURN sql given * an entity definition * updates to be run on unique key(s) collision

When left as Nothing, we find the unique key from entity def before * trying to fetch an entity by said key * perform an update when result found, else issue an insert * return new entity from db

Since: 2.6

connPutManySql :: Maybe (EntityDef -> Int -> Text)

Some databases support performing bulk UPSERT, specifically "insert or replace many records" in a single call.

This field when set, given * an entity definition * number of records to be inserted should produce a PUT MANY sql with placeholders for records

When left as Nothing, we default to using defaultPutMany.

Since: 2.8.1

connStmtMap :: IORef (Map Text Statement)

connClose :: IO ()

connMigrateSql :: [EntityDef] -> (Text -> IO Statement) -> EntityDef -> IO (Either [Text] [(Bool, Text)])

connBegin :: (Text -> IO Statement) -> Maybe IsolationLevel -> IO ()

connCommit :: (Text -> IO Statement) -> IO ()

connRollback :: (Text -> IO Statement) -> IO ()

connEscapeName :: DBName -> Text

connNoLimit :: Text

connRDBMS :: Text

connLimitOffset :: (Int, Int) -> Bool -> Text -> Text

connLogFunc :: LogFunc

connMaxParams :: Maybe Int

Some databases (probably only Sqlite) have a limit on how many question-mark parameters may be used in a statement

Since: 2.6.1

connRepsertManySql :: Maybe (EntityDef -> Int -> Text)

Some databases support performing bulk an atomic+bulk INSERT where constraint conflicting entities can replace existing entities.

This field when set, given * an entity definition * number of records to be inserted should produce a INSERT sql with placeholders for primary+record fields

When left as Nothing, we default to using defaultRepsertMany.

Since: 2.9.0

type SqlBackendCanRead backend = (BackendCompatible SqlBackend backend, PersistQueryRead backend, PersistStoreRead backend, PersistUniqueRead backend)

A constraint synonym which witnesses that a backend is SQL and can run read queries.

type SqlBackendCanWrite backend = (SqlBackendCanRead backend, PersistQueryWrite backend, PersistStoreWrite backend, PersistUniqueWrite backend)

A constraint synonym which witnesses that a backend is SQL and can run read and write queries.

type SqlReadT m a = forall backend. SqlBackendCanRead backend => ReaderT backend m a

Like SqlPersistT but compatible with any SQL backend which can handle read queries.

type SqlWriteT m a = forall backend. SqlBackendCanWrite backend => ReaderT backend m a

Like SqlPersistT but compatible with any SQL backend which can handle read and write queries.

type IsSqlBackend backend = (IsPersistBackend backend, BaseBackend backend ~ SqlBackend)

A backend which is a wrapper around SqlBackend.